refactor: 에러 코드 exceptions 패키지로 이동
