+++ src/main/java/roomescape/exceptions/ErrorCode.java
@@ -1,4 +1,4 @@
-package roomescape.errors;
+package roomescape.exceptions;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
